+++ src/store/actions/creators/update-favorite-count.creator.js
@@ -1,0 +1,8 @@
+import UPDATE_FAVORITE_COUNT from '../types/update-favorite-count.type';
+
+const updateFavoriteCount = payload => ({
+  type: UPDATE_FAVORITE_COUNT,
+  payload,
+});
+
+export default updateFavoriteCount;