+++ src/store/get-initial-state.js
@@ -1,5 +1,6 @@
 import isMobile from '@config/is-mobile';
-import getOnlineStatus from './get-online-status';
+import getOnlineStatus from '@utils/get-online-status';
+import objToMap from '@utils/obj-to-map';
 import getSortedFilteredPieceIds from './get-sorted-filtered-piece-ids';
 import { version } from '../../package.json';
 
@@ -34,6 +35,14 @@ const getInitialState = storedState => {
     cachedPieceIds: isNewVersion
       ? new Set(storedState.cachedPieceIds)
       : new Set(),
+    globalPlayTime:
+      typeof storedState.globalPlayTime === 'object'
+        ? objToMap(storedState.globalPlayTime)
+        : new Map(),
+    favoriteCount:
+      typeof storedState.favoriteCount === 'object'
+        ? objToMap(storedState.favoriteCount)
+        : new Map(),
   });
 
   // must be done after previous step (converting favorites to Set)