+++ src/utils/obj-to-map.js
@@ -1,0 +1,7 @@
+const objToMap = o =>
+  Reflect.ownKeys(o).reduce((map, key) => {
+    map.set(key, o[key]);
+    return map;
+  }, new Map());
+
+export default objToMap;